// src/app/models/perfil.model.ts
export class PerfilModel {

    id?:string = '';
    fecha: string = '';
    socio: string = '';
    cliente_prospecto: string = ''
    cod: string = '';
    consultor: string = '';
    sucursal: string = '';
    giro: string = '';
    rubro: string = '';
    totalEmpleados: string = '';
    diasLaborales: string = '';
    produccion: string = '';
    costoProduccion: string = '';

    costoQuimico: string = '';
    cuadrillaSanidad: string = '';
    costoMO: string = '';
    turnosLaborales: string = '';
    accesorias: string = '';

    presupuestoAsignado: string = '';
    pve: number = 0;
    desicionCompra: string = '';
    preocupacionNecesidad: string = '';
    referenciaCliente: string = '';
    marcasConocidad: string = '';
    socioCliente: string = '';
    procesoLimpieza: string = '';
    proveedorLimpieza: string = '';
    sistemaInocuidadCalidad: any[] = [];
    otro: string = '';
    observaciones: string = '';

    perfil: string = '';

    estrategia: string = '';

    pms: any[] = [];
}